+++ functions/src/image/skip-image.test.ts
