+++ functions/src/image/skip-image.ts
